requirements/common.txt
#Django
django>=4.0,<4.1
djangorestframework>=3.13.0
djangorestframework-camel-case<=1.3.0
psycopg2-binary>=2.8
django-cors-headers>=3.11.0
pyyaml<7,>=6.0
pyjwt
django-filter==22.1


